+++ just-swim-app/.eslintrc.json
@@ -1,5 +1,5 @@
 {
-  "extends": ["next/babel", "next/core-web-vitals"],
+  "extends": ["next/core-web-vitals"],
   "rules": {
     "no-console": "warn",
     "no-unused-vars": "off"
